Added Photon Room script, added multiplayer scene
I have sucessfully made it video 6, we are able to connect with one player hosting the game and the other joining with them. This is a quick start, everything that is commented out is for delayed start. As of right now it all works just need to add spawn points for players to spawn in. This is up to video 5.