Added delay start room controller w/ gameobject, added waiting room controller, w/ gameobject,
this allows the player to press the battle button, then loads the playe into the waiting room. there is something missing in it that is not counting the player when loaded into the waiting room. I added the AARMController which counts the player. please look over this and let me know what you find.